requirements.txt
alembic==0.8.8
amqp==2.2.2
apipkg==1.4
Babel==0.9.6
beautifulsoup4==4.4.1
billiard==3.5.0.3
bleach==1.5.0
blinker==1.4
celery==4.1.0
chardet==2.2.1
click==6.6
coverage==4.2
cycler==0.10.0
DBUtils==1.1
devpi-common==3.0.1
devpi-server==4.1.1
dominate==2.3.0
execnet==1.4.1
flake8==2.0
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Celery==2.4.3
Flask-HTTPAuth==3.2.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.1
Flask-Moment==0.5.1
Flask-PageDown==0.1.3
flask-paginate==0.4.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
Flask-WTF==0.13.1
ForgeryPy==0.1
futures==3.0.3
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.8
kombu==4.1.0
Mako==1.0.6
Markdown==2.6.7
markerlib==0.6.0
MarkupSafe==0.23
mccabe==0.5.2
msgpack-python==0.4.6
numpy==1.11.0
PasteDeploy==1.5.2
pep8==1.7.0
pep8-naming==0.4.1
pluggy==0.4.0
psutil==5.2.2
py==1.4.31
pycrypto==2.6.1
pyflakes==1.3.0
Pygments==2.1.3
pymongo==3.0.3
PyMySQL==0.7.11
pyparsing==2.1.4
pyramid==1.7.3
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.4
PyYAML==3.11
pyzmq==14.5.0
redis==2.10.5
repoze.lru==0.6
requests==2.12.1
salt==2017.7.2
selenium==3.0.1
six==1.10.0
SQLAlchemy==1.1.4
tornado==4.2.1
translationstring==1.3
urllib3==1.10.4
venusian==1.0
vine==1.1.4
visitor==0.1.3
waitress==0.9.0
WebOb==1.6.3
Werkzeug==0.11.11
WTForms==2.1
zope.deprecation==4.2.0
zope.interface==4.3.2
